// packages/codegen/src/output/database/src/mongoose/interfaces/iProjectTemplateCreateInput.ts
// THIS CODE WAS AUTOMATICALLY GENERATED
import {databaseTypes} from 'types';
import {Types as mongooseTypes} from 'mongoose';

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface IProjectTemplateCreateInput
  extends Omit<databaseTypes.IProjectTemplate, '_id' | 'createdAt' | 'updatedAt'> {}